fix: address review comments (rename files, fix logging)
