+++ src/core/application/use-cases/auth/checkSsoUseCase.ts
