+++ src/core/domain/organizationParameters/types/ssoConfig.type.ts
